Better organization of Websocket modules
